requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docker==4.2.0
gitdb==4.0.5
GitPython==3.1.2
idna==2.9
requests==2.23.0
six==1.15.0
smmap==3.0.4
urllib3==1.25.9
websocket-client==0.57.0
